docs/requirements.txt
pyyaml
markdown
pymongo 
dnspython 
python-dotenv 
beautifulsoup4
openai
